+++ CHANGELOG.md
@@ -48,6 +48,9 @@ with the `YY.MINOR.MICRO` scheme.
 - SQL supports `LIKE` operator
 - SQL supports `ILIKE` operator
 - SQL supports `lower` and `upper` string functions
+- Execute option `sql_vdbe_max_steps` was renamed to
+`vdbe_max_steps`
+
 
 ## [24.5.1] - 2024-09-04